sea_battle/sea_battle_app/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-auth==0.1.9
channels==3.0.4
channels-redis==3.3.1
python-dotenv==0.19.2
psycopg2-binary==2.9.2